dashboard/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
deltalake>=0.15.0
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
